Started fixing bugs related to macOS namely handling %, (null) and (nil) and some floating point exceptions

+++ includes/ft_ftoa.h
@@ -6,7 +6,7 @@
 /*   By: dda-silv <dda-silv@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/21 08:32:33 by dda-silv          #+#    #+#             */
-/*   Updated: 2021/01/23 08:48:36 by dda-silv         ###   ########.fr       */
+/*   Updated: 2021/02/14 19:55:49 by dda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 # include "libft.h"
 
-static char		*round_after_decimal(double nb, int length, char *str_nb);
+static char		*round_after_decimal(long double nb, int length, char *str_nb);
 static char		*round_before_decimal(char *str_nb, int length);
 static int		get_new_length(char *str_nb, int length);
 static void		copy_after_decimal(char *str_nb,
